Modified

+++ train.py
@@ -68,8 +68,8 @@ def sequence_loss(flow_preds, flow_gt, valid, gamma=0.8, max_flow=MAX_FLOW):
         ref, sparse_flow = flow_preds[1][i]
         scale = torch.tensor((I_W, I_H), dtype=torch.float32).view(1, 1, 2).to(sparse_flow.device)
         flatten_gt = flow_gt.flatten(2).permute(0, 2, 1)
-        ceil_coords = torch.ceil(ref * scale).int()
-        floor_coords = torch.floor(ref * scale).int()
+        ceil_coords = torch.ceil(ref * scale).long()
+        floor_coords = torch.floor(ref * scale).long()
         sparse_gt = flatten_gt[torch.arange(bs), :,
                     floor_coords[torch.arange(bs), :, 1] * floor_coords[torch.arange(bs), :, 0]] * ref.frac() + \
                     flatten_gt[torch.arange(bs), :,
